+++ scripts/utils/api.js
@@ -1,4 +1,4 @@
-import { BASE_URL, MINER_UPGRADE_URL, MINER_BUY_URL } from "../config.js";
+import { BASE_URL, MINER_UPGRADE_URL, MINER_BUY_URL, DISCOUNT_URL, MAINTENANCE_STATE_URL } from "../config.js";
 
 /**
  * Generic function to fetch data from an API endpoint with a bearer token.
@@ -27,13 +27,18 @@ export async function fetchData(skip, bearerToken) {
         if (response.status === 401) {
             throw new Error("Unauthorized. Please ensure you are logged in.");
         }
-        throw new Error(`API request failed: ${response.statusText}`);
+        throw new Error(`API gmt wallet request failed: ${response.statusText}`);
     }
 
     return await response.json();
 }
 
-
+/**
+ * Generic function to fetch bought nft from an API endpoint with a bearer token.
+ * @param {int} skip - For paging - skip at least 'skip' values.
+ * @param {string} bearerToken - Bearer token for authorization.
+ * @returns {Promise<Object>} The response data.
+ */
 export async function fetchNFT(skip, bearerToken) {
     const headers = {
         Authorization: `Bearer ${bearerToken}`,
@@ -55,12 +60,18 @@ export async function fetchNFT(skip, bearerToken) {
         if (response.status === 401) {
             throw new Error("Unauthorized. Please ensure you are logged in.");
         }
-        throw new Error(`API request failed: ${response.statusText}`);
+        throw new Error(`API nft order request failed: ${response.statusText}`);
     }
 
     return await response.json();
 }
 
+/**
+ * Generic function to fetch nft upgrades from an API endpoint with a bearer token.
+ * @param {int} skip - For paging - skip at least 'skip' values.
+ * @param {string} bearerToken - Bearer token for authorization.
+ * @returns {Promise<Object>} The response data.
+ */
 export async function fetchUpgradesNFT(skip, bearerToken) {
     const headers = {
         Authorization: `Bearer ${bearerToken}`,
@@ -92,7 +103,65 @@ export async function fetchUpgradesNFT(skip, bearerToken) {
         if (response.status === 401) {
             throw new Error("Unauthorized. Please ensure you are logged in.");
         }
-        throw new Error(`API request failed: ${response.statusText}`);
+        throw new Error(`API nft upgrade request failed: ${response.statusText}`);
+    }
+
+    return await response.json();
+}
+
+/**
+ * Generic function to fetch discount information from an API endpoint with a bearer token.
+ * @param {string} bearerToken - Bearer token for authorization.
+ * @returns {Promise<Object>} The response data.
+ */
+export async function fetchDiscount(bearerToken) {
+    const headers = {
+        Authorization: `Bearer ${bearerToken}`,
+        "Content-Type": "application/json",
+    };
+
+    const body = {};
+
+    const response = await fetch(DISCOUNT_URL, {
+        method: "POST",
+        headers: headers,
+        body: JSON.stringify(body),
+    });
+
+    if (!response.ok) {
+        if (response.status === 401) {
+            throw new Error("Unauthorized. Please ensure you are logged in.");
+        }
+        throw new Error(`API discount request failed: ${response.statusText}`);
+    }
+
+    return await response.json();
+}
+
+/**
+ * Generic function to fetch the maintenance state information from an API endpoint with a bearer token.
+ * @param {string} bearerToken - Bearer token for authorization.
+ * @returns {Promise<Object>} The response data.
+ */
+export async function fetchMaintenanceState(bearerToken) {
+    const headers = {
+        Authorization: `Bearer ${bearerToken}`,
+        "Content-Type": "application/json",
+    };
+
+    const body = {};
+
+    const response = await fetch(MAINTENANCE_STATE_URL, {
+        method: "POST",
+        headers: headers,
+        body: JSON.stringify(body),
+    });
+
+    if (!response.ok) {
+        if (response.status === 401) {
+            throw new Error("Unauthorized. Please ensure you are logged in.");
+        }
+        throw new Error(`API maintenance state request failed: ${response.statusText}`);
     }
 
     return await response.json();